new and improved (yes this time really)11

+++ Wep lommeregner/Models/ReturnPointAndQuestion.cs
@@ -2,7 +2,7 @@
 {
     public class ReturnPointAndQuestion
     {
-        public decimal Points { get; set; }
+        public double Points { get; set; }
         public Question Question { get; set; }
         
     }
